backend/pyproject.toml
[project]
name = "start-template-fastapi"
version = "0.1.0"
description = ""
authors = [
    {name = "vladislav1029",email = "vlad1090904@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11, <4.0" # иначе ищит поддержку аж до 4 версии. 
dependencies = [
    "fastapi (>=0.116.2,<0.117.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "pyjwt[crypto] (>=2.10.1,<3.0.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "structlog (>=25.4.0,<26.0.0)",
    "pydantic[email] (>=2.12.0,<3.0.0)",
    "text-unidecode (>=1.3,<2.0)",
    "minio (>=7.2.18,<8.0.0)", # INFO: либа sync блокает event loop, есть конечно (miniopy-async)
    "python-multipart (>=0.0.20,<0.0.21)", # INFO: для нормальных форм в swagger
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "bcrypt (==4.0.1)",# INFO: Так как это стабильная версия
    "cryptography (>=46.0.3,<47.0.0)",# INFO: для RS256
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

